+++ app/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -16,6 +16,7 @@ package org.hyperledger.besu.controller;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import org.hyperledger.besu.chainimport.BlockHeadersCachePreload;
 import org.hyperledger.besu.components.BesuComponent;
 import org.hyperledger.besu.config.CheckpointConfigOptions;
 import org.hyperledger.besu.config.GenesisConfig;
@@ -209,6 +210,7 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
 
   private int numberOfBlocksToCache = 0;
   private int numberOfBlockHeadersToCache = 0;
+  private boolean isCacheLastBlockHeadersPreloadEnabled;
 
   /** whether parallel transaction processing is enabled or not */
   protected boolean isParallelTxProcessingEnabled;
@@ -515,6 +517,19 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
   }
 
   /**
+   * Sets whether the block header cache should be preloaded.
+   *
+   * @param isCacheLastBlockHeadersPreloadEnabled {@code true} to enable preloading of the block
+   *     header cache, {@code false} to disable it
+   * @return this builder instance
+   */
+  public BesuControllerBuilder isCacheLastBlockHeadersPreloadEnabled(
+      final Boolean isCacheLastBlockHeadersPreloadEnabled) {
+    this.isCacheLastBlockHeadersPreloadEnabled = isCacheLastBlockHeadersPreloadEnabled;
+    return this;
+  }
+
+  /**
    * sets the networkConfiguration in the builder
    *
    * @param networkingConfiguration the networking config
@@ -623,6 +638,13 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
             protocolSchedule,
             codeCache);
 
+    final EthScheduler scheduler =
+        new EthScheduler(
+            syncConfig.getDownloaderParallelism(),
+            syncConfig.getTransactionsParallelism(),
+            syncConfig.getComputationParallelism(),
+            metricsSystem);
+
     final MutableBlockchain blockchain =
         DefaultBlockchain.createMutable(
             genesisState.getBlock(),
@@ -632,6 +654,13 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
             dataDirectory.toString(),
             numberOfBlocksToCache,
             numberOfBlockHeadersToCache);
+
+    if (isCacheLastBlockHeadersPreloadEnabled && numberOfBlockHeadersToCache > 0) {
+      LOG.info(
+          "--cache-last-block-headers and --cache-last-block-headers-preload-enabled are enabled, start preloading block headers cache");
+      preloadBlockHeaderCache(blockchain, scheduler);
+    }
+
     final BonsaiCachedMerkleTrieLoader bonsaiCachedMerkleTrieLoader =
         besuComponent
             .map(BesuComponent::getCachedMerkleTrieLoader)
@@ -695,13 +724,6 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
 
     final EthMessages ethMessages = new EthMessages();
     final EthMessages snapMessages = new EthMessages();
-
-    final EthScheduler scheduler =
-        new EthScheduler(
-            syncConfig.getDownloaderParallelism(),
-            syncConfig.getTransactionsParallelism(),
-            syncConfig.getComputationParallelism(),
-            metricsSystem);
 
     Optional<Checkpoint> checkpoint = Optional.empty();
     if (genesisConfigOptions.getCheckpointOptions().isValid()) {
@@ -879,6 +901,33 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
         storageProvider,
         dataStorageConfiguration,
         transactionSimulator);
+  }
+
+  private void preloadBlockHeaderCache(
+      final MutableBlockchain blockchain, final EthScheduler scheduler) {
+    final BlockHeadersCachePreload blockHeaderCachePreload =
+        new BlockHeadersCachePreload(blockchain, scheduler, numberOfBlockHeadersToCache);
+    long startTime = System.nanoTime();
+    blockHeaderCachePreload
+        .preloadCache()
+        .thenRun(
+            () -> {
+              long duration = System.nanoTime() - startTime;
+              LOG.info(
+                  "Preloading {} block headers to the cache finished in {} ms",
+                  numberOfBlockHeadersToCache,
+                  duration / 1_000_000);
+            })
+        .exceptionally(
+            throwable -> {
+              long duration = System.nanoTime() - startTime;
+              LOG.error(
+                  "Preloading {} block headers to the cache failed after {} ms",
+                  numberOfBlockHeadersToCache,
+                  duration / 1_000_000,
+                  throwable);
+              return null;
+            });
   }
 
   private GenesisState getGenesisState(